Implement grey text pills
Also reorganized storybook to clean up the quickly-growing file.

Closes #35.
