+++ storybook/storybook.js
@@ -3,10 +3,10 @@ import React, { Component } from 'react';
 import { AppRegistry } from 'react-native';
 import { getStorybookUI, configure } from '@storybook/react-native';
 
+import { loadStories } from './storyLoader';
+
 // import stories
-configure(() => {
-  require('./stories');
-}, module);
+configure(loadStories, module);
 
 // This assumes that storybook is running on the same host as your RN packager,
 // to set manually use, e.g. host: 'localhost' option
